fix integer arithmetic; prevent possible overflow

+++ core/src/main/java/com/microsoft/applicationinsights/internal/channel/common/SenderThreadsBackOffManager.java
@@ -83,8 +83,8 @@ final class SenderThreadsBackOffManager extends ThreadLocal<SenderThreadLocalBac
 
     @Override
     protected SenderThreadLocalBackOffData initialValue() {
-        int addSeconds = threadsSecondsDifference.incrementAndGet();
-        senderThreadLocalData = new SenderThreadLocalBackOffData(backOffTimeoutsInMilliseconds, addSeconds * 1000);
+        long addSeconds = threadsSecondsDifference.incrementAndGet();
+        senderThreadLocalData = new SenderThreadLocalBackOffData(backOffTimeoutsInMilliseconds, addSeconds * 1000L);
         registerSenderData(senderThreadLocalData);
         return senderThreadLocalData;
     }
